move: MemberRegistRequestDto 경로 변경
